backend/requirements.txt
alpaca-py>=0.31.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pymongo>=4.8.0
pandas>=2.2.0